requirements.txt
pdfplumber
python-dotenv
tiktoken
tqdm
faiss-cpu
openai
gradio
numpy
openai-agents
fastapi
uvicorn[standard]
pydantic